App finalizada
Adds likes and show fav games list. Also games CRUD works correctly.

+++ handlers/videojuego/delete.py
@@ -3,7 +3,7 @@ from time import sleep
 
 import webapp2
 from google.appengine.api import images
-from PIL import Image
+from webapp2_extras.users import users
 from webapp2_extras import jinja2
 
 
@@ -14,11 +14,18 @@ from model.videojuego import Pegi
 
 class EliminaVideojuegoHandler(webapp2.RequestHandler):
     def get(self):
-        videojuego = Videojuego.recupera(self.request)
-        videojuego.key.delete()
-        sleep(1)
+        usr = users.get_current_user()
 
-        return self.redirect("/")
+        if usr:
+            videojuego = Videojuego.recupera(self.request)
+            videojuego.key.delete()
+            sleep(1)
+
+            return self.redirect("/")
+        else:
+            print("No esta loggeado en delete")
+            return self.redirect("/")
+
 
 
 app = webapp2.WSGIApplication([
